Add a removing test and search a row in a job table. Create a Table and Table Card classes.
