+++ pom.xml
@@ -19,7 +19,7 @@
         <checkstyle-maven-plugin.version>3.6.0</checkstyle-maven-plugin.version>
         <maven-surefire-plugin.version>3.5.3</maven-surefire-plugin.version>
         <spotless.version>2.44.3</spotless.version>
-        <selenium.version>4.33.0</selenium.version>
+        <selenium.version>4.34.0</selenium.version>
         <javafaker.version>1.0.2</javafaker.version>
         <lombok.version>1.18.38</lombok.version>
     </properties>
