+++ src/main/java/eu/senla/commonElement/Sidepanel.java
@@ -1,19 +1,19 @@
 package eu.senla.commonElement;
 
 import eu.senla.enums.SidepanelMenu;
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-import eu.senla.pageObject.login.auth.DashboardPage;
-import eu.senla.pageObject.login.auth.admin.AdminPage;
-import eu.senla.pageObject.login.auth.buzz.BuzzPage;
-import eu.senla.pageObject.login.auth.claim.ClaimPage;
-import eu.senla.pageObject.login.auth.directory.DirectoryPage;
-import eu.senla.pageObject.login.auth.leave.LeavePage;
-import eu.senla.pageObject.login.auth.maintenance.MaintenancePage;
-import eu.senla.pageObject.login.auth.myInfo.MyInfoPage;
-import eu.senla.pageObject.login.auth.performance.PerformancePage;
-import eu.senla.pageObject.login.auth.pim.PimPage;
-import eu.senla.pageObject.login.auth.recruitment.RecruitmentPage;
-import eu.senla.pageObject.login.auth.time.TimePage;
+import eu.senla.pageObject.auth.BaseAuthPage;
+import eu.senla.pageObject.auth.DashboardPage;
+import eu.senla.pageObject.auth.admin.AdminPage;
+import eu.senla.pageObject.auth.buzz.BuzzPage;
+import eu.senla.pageObject.auth.claim.ClaimPage;
+import eu.senla.pageObject.auth.directory.DirectoryPage;
+import eu.senla.pageObject.auth.leave.LeavePage;
+import eu.senla.pageObject.auth.maintenance.MaintenancePage;
+import eu.senla.pageObject.auth.myInfo.MyInfoPage;
+import eu.senla.pageObject.auth.performance.PerformancePage;
+import eu.senla.pageObject.auth.pim.PimPage;
+import eu.senla.pageObject.auth.recruitment.RecruitmentPage;
+import eu.senla.pageObject.auth.time.TimePage;
 import eu.senla.utilities.Wait;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
