+++ src/main/java/eu/senla/pageObject/auth/buzz/BuzzPage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.buzz;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class BuzzPage extends BaseAuthPage {
+}
