+++ src/main/java/eu/senla/pageObject/auth/claim/ClaimPage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.claim;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class ClaimPage extends BaseAuthPage {
+}
