+++ src/main/java/eu/senla/pageObject/auth/directory/DirectoryPage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.directory;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class DirectoryPage extends BaseAuthPage {
+}
