+++ src/main/java/eu/senla/pageObject/auth/leave/LeavePage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.leave;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class LeavePage extends BaseAuthPage {
+}
