+++ src/main/java/eu/senla/pageObject/auth/maintenance/MaintenancePage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.maintenance;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class MaintenancePage extends BaseAuthPage {
+}
