+++ src/main/java/eu/senla/pageObject/auth/myInfo/MyInfoPage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.myInfo;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class MyInfoPage extends BaseAuthPage {
+}
