+++ src/main/java/eu/senla/pageObject/auth/performance/PerformancePage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.performance;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class PerformancePage extends BaseAuthPage {
+}
