+++ src/main/java/eu/senla/pageObject/auth/pim/PersonalDetailPage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.pim;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class PersonalDetailPage extends BaseAuthPage {
+}
