+++ src/main/java/eu/senla/pageObject/auth/recruitment/RecruitmentPage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.recruitment;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class RecruitmentPage extends BaseAuthPage {
+}
