+++ src/main/java/eu/senla/pageObject/auth/time/TimePage.java
@@ -1,0 +1,6 @@
+package eu.senla.pageObject.auth.time;
+
+import eu.senla.pageObject.auth.BaseAuthPage;
+
+public class TimePage extends BaseAuthPage {
+}
