--- src/main/java/eu/senla/pageObject/login/auth/DashboardPage.java
@@ -1,13 +1,0 @@
-package eu.senla.pageObject.login.auth;
-
-import eu.senla.utilities.Wait;
-import org.openqa.selenium.By;
-
-public class DashboardPage extends BaseAuthPage {
-
-    private final By pageHeaderBy = By.className("oxd-text--h6");
-
-    public final String getPageHeader() {
-        return Wait.waitVisibilityOfElementLocated(pageHeaderBy).getText();
-    }
-}
