--- src/main/java/eu/senla/pageObject/login/auth/buzz/BuzzPage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.buzz;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class BuzzPage extends BaseAuthPage {
-}
