--- src/main/java/eu/senla/pageObject/login/auth/claim/ClaimPage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.claim;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class ClaimPage extends BaseAuthPage {
-}
