--- src/main/java/eu/senla/pageObject/login/auth/directory/DirectoryPage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.directory;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class DirectoryPage extends BaseAuthPage {
-}
