--- src/main/java/eu/senla/pageObject/login/auth/leave/LeavePage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.leave;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class LeavePage extends BaseAuthPage {
-}
