--- src/main/java/eu/senla/pageObject/login/auth/maintenance/MaintenancePage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.maintenance;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class MaintenancePage extends BaseAuthPage {
-}
