--- src/main/java/eu/senla/pageObject/login/auth/myInfo/MyInfoPage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.myInfo;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class MyInfoPage extends BaseAuthPage {
-}
