--- src/main/java/eu/senla/pageObject/login/auth/performance/PerformancePage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.performance;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class PerformancePage extends BaseAuthPage {
-}
