--- src/main/java/eu/senla/pageObject/login/auth/pim/PersonalDetailPage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.pim;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class PersonalDetailPage extends BaseAuthPage {
-}
