--- src/main/java/eu/senla/pageObject/login/auth/recruitment/RecruitmentPage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.recruitment;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class RecruitmentPage extends BaseAuthPage {
-}
