--- src/main/java/eu/senla/pageObject/login/auth/time/TimePage.java
@@ -1,6 +1,0 @@
-package eu.senla.pageObject.login.auth.time;
-
-import eu.senla.pageObject.login.auth.BaseAuthPage;
-
-public class TimePage extends BaseAuthPage {
-}
