+++ src/main/java/eu/senla/webElement/Input.java
@@ -1,12 +1,14 @@
 package eu.senla.webElement;
 
+import org.openqa.selenium.By;
+
 public class Input extends BaseCustomElement {
 
     public Input(final String name) {
         super(name);
     }
 
-    final String getXpathExpression() {
-        return "//form//label[text()='" + super.getName() + "']/../../div/input";
+    protected final By getBy() {
+        return By.xpath("//form//label[text()='" + super.getName() + "']/../../div/input");
     }
 }
