+++ src/main/java/eu/senla/webElement/table/TableCard.java
@@ -1,0 +1,22 @@
+package eu.senla.webElement.table;
+
+import eu.senla.webElement.BaseCustomElement;
+import org.openqa.selenium.By;
+
+public abstract class TableCard extends BaseCustomElement {
+
+    public TableCard(final String name) {
+        super(name);
+    }
+
+    protected final String getRowXpath() {
+        return "//div[@class='oxd-table-card']//div[text()='"
+                + super.getName()
+                + "']/ancestor::div[contains(@class,'row')]";
+    }
+
+    protected final By getBy() {
+        return By.xpath(getRowXpath());
+    }
+
+}
